refactor: move displays to package views
Additionally, move ClubManagementCLI to package main. How do I write better logs?
