+++ src/main/ClubManagementSystem.java
@@ -1,7 +1,5 @@
 package main;
 
-import main.views.cli.ClubManagementCLI;
-
 /**
  * Class representing a Health and Fitness Club Management System.
  */
