--- src/main/views/cli/commands/Command.java
@@ -1,3 +1,0 @@
-package main.views.cli.commands;
-
-public interface Command {}
