+++ src/main/views/commands/Command.java
@@ -1,0 +1,3 @@
+package main.views.commands;
+
+public interface Command {}
